Sorting refactoring & comments/docs
