+++ slicker-postgres/src/main/scala/io/slicker/postgres/RecordTable.scala
@@ -1,6 +1,6 @@
 package io.slicker.postgres
 
-import io.slicker.core.order.Order
+import io.slicker.core.sort.Sort
 import io.slicker.postgres.PostgresDriver.api._
 
 /**
@@ -28,6 +28,10 @@ trait RecordTable[Id, Business, Database, T <: TableWithId[Id, Database]] {
     */
   def toBusiness(database: Database): Business
 
-  def order: Order[T] = Order.empty[T]
+  /**
+    * [[Sort]] object that will return [[slick.lifted.Ordered]]
+    * for some input string as field name to sort by
+    */
+  def order: Sort[T] = Sort.empty[T]
 
 }
