+++ slicker-postgres/src/test/scala/io/slicker/postgres/PostgreSQLRepositorySpec.scala
@@ -1,6 +1,6 @@
 package io.slicker.postgres
 
-import io.slicker.core.{PageRequest, Sort, SortDirection}
+import io.slicker.core.{Fields, PageRequest, SortDirection}
 import io.slicker.postgres.PostgresDriver.api._
 import io.slicker.postgres.models.{User, UserTable, UsersRepository}
 import org.scalatest.{BeforeAndAfter, BeforeAndAfterAll, FlatSpec, Matchers}
@@ -45,7 +45,7 @@ class PostgreSQLRepositorySpec extends FlatSpec with Matchers with AwaitHelper w
 
   it should "save multiple entities and find them in descending order" in {
     val users = userRepo.save(Seq(User(None, "name"), User(None, "name2"))).run.await
-    val sort = Sort(Seq("name" -> SortDirection.Desc))
+    val sort = Fields(Seq("name" -> SortDirection.Desc))
     userRepo.findAll(PageRequest(sort)).run.await shouldBe users.sortBy(_.name).reverse
   }
