+++ slicker-postgres/src/test/scala/io/slicker/postgres/models/UserTable.scala
@@ -1,6 +1,6 @@
 package io.slicker.postgres.models
 
-import io.slicker.core.order.Order
+import io.slicker.core.sort.Sort
 import io.slicker.postgres.PostgresDriver.api._
 import io.slicker.postgres.{SimpleRecordTable, TableWithId}
 import slick.lifted.ProvenShape
@@ -11,7 +11,7 @@ class UserTable extends SimpleRecordTable[Long, User, Users] {
 
   override val tableQuery: TableQuery[Users] = TableQuery[Users]
 
-  override def order: Order[Users] = Order.full[Users]
+  override def order: Sort[Users] = Sort.auto[Users]
 
 }
